Improved fitting, tooltips, MAGI ground work
Our fitting always makes sure we stay above the target estate.
We can fit for retirement age and then subsequently for expense which works nicely.
Added tooltips to a lot of the entry fields.
Put a lot of stuff in place to be able to evaluate and optimize for a maximum MAGI so we can start getting more complicated about draw down.